requirements.txt
fastapi
uvicorn
aiosqlite
pyjwt
bcrypt
httpx
pytest
pytest-asyncio
